🚚 rename recoil state folder
